Move aggregate_lock_specs to its own module
